Add Logger plugin Used in printing out things on the console in a clear and organized manner
Add Shared preferences
To be used for local storage in phone

Build Dummy Data class file
To store datas like strings, boolean, int values and the rest to be used across app

Build Custom Button State
This class takes in a text, and an already created enum named ButtonState. This ButtonState is of three types namely;
idle, disabled, and loading.